multi-tenant-configuration/requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
requests==2.25.0
requests-toolbelt==0.9.1
urllib3==1.26.2

pyyaml